feat(ci/docker): add GitHub Actions and Dockerfile
This commit introduces two major changes. Firstly, a new GitHub Actions
workflow is added for the backend. This workflow runs tests and checks
the formatting of the code using ESLint whenever a pull request is made
to the main branch.

Secondly, a new Dockerfile is added to the backend. This Dockerfile
describes how to build a Docker image for the backend. It uses a
multi-stage build process to create a smaller final image.